Рефакторинг
Исправил нэйминг
Слайдер переместил в отдельную папку вместе с его стилями (работает)
