+++ src/components/feedback/feedback.components.jsx
@@ -1,4 +1,4 @@
-import './feedback.component.css';
+import './feedback.components.css';
 import "yup-phone";
 import React from 'react';
 import { Formik, Form, Field } from 'formik';
