+++ src/components/header/header.components.css
@@ -1,4 +1,4 @@
-/* header */
+
 .menu-item {
   flex-basis: 20%;
   font-size: 24px;
