--- src/styles.css
@@ -1,36 +1,0 @@
-.App {
-  font-family: sans-serif;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  height: 100vh;
-}
-
-.seperator {
-  width: 30%;
-}
-
-.controls-wrapper {
-  width: 100%;
-}
-
-.carousel-wrapper {
-  width: 100%;
-}
-
-.rec-carousel-item:focus {
-  outline: none;
-  box-shadow: inset 0 0 1px 0px gray;
-}
-.font {
-  font-style: italic;
-  font-size: 30px;
-}
-/*.hyperlinks {
-  width: "100%";
-  display: "flex";
-  flex-direction:"column";
-  align-items: "center";
-  margin-left: "10px";
-}*/
